Move RenderEvent to RenderChatTimestampEvent
